requirements.txt
asgiref==3.2.7
awswrangler==1.0.4
beautifulsoup4==4.9.0
boto3==1.12.44
botocore==1.15.44
certifi==2020.4.5.1
chardet==3.0.4
cycler==0.10.0
defusedxml==0.6.0
Django==3.0.5
django-allauth==0.41.0
django-bootstrap4==1.1.1
django-pandas==0.6.1
django-rest==0.0.1
django-rest-framework==0.1.0
django-widget-tweaks==1.4.8
djangorestframework==3.11.0
docutils==0.15.2
fsspec==0.7.3
idna==2.9
importlib-metadata==1.6.0
inflect==4.1.0
jmespath==0.9.5
joblib==0.14.1
kiwisolver==1.2.0
matplotlib==3.2.1
mplfinance==0.12.3a3
numpy==1.18.2
oauthlib==3.1.0
pandas==1.0.3
Pillow==7.1.1
plotly==4.6.0
psycopg2-binary==2.8.5
pyarrow==0.16.0
PyMySQL==0.9.3
pyparsing==2.4.7
PyQt5==5.14.2
PyQt5-sip==12.7.2
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
retrying==1.3.3
s3fs==0.4.2
s3transfer==0.3.3
scikit-learn==0.22.2.post1
scipy==1.4.1
seaborn==0.10.0
six==1.14.0
sklearn==0.0
soupsieve==2.0
sqlacodegen==2.1.0
SQLAlchemy==1.3.13
sqlalchemy-redshift==0.7.7
sqlparse==0.3.1
urllib3==1.25.8
zipp==3.1.0
